import flattenObject function
