+++ apps/admin-server/src/pages/projects/[project]/resources/index.tsx
@@ -10,6 +10,7 @@ import { RemoveResourceDialog } from '@/components/dialog-resource-remove';
 import { toast } from 'react-hot-toast';
 import { sortTable, searchTable } from '@/components/ui/sortTable';
 import * as XLSX from 'xlsx';
+import flattenObject from "@/lib/export-helpers/flattenObject";
 
 export default function ProjectResources() {
   const router = useRouter();
@@ -17,28 +18,13 @@ export default function ProjectResources() {
   const { data, error, isLoading, remove } = useResources(project as string);
 
   const exportData = (data: any[], fileName: string) => {
-    const flattenObject = (obj: any, parent: string = '', res: any = {}) => {
-      for (let key in obj) {
-        const propName = parent ? `${parent}.${key}` : key;
-        if (typeof obj[key] === 'object' && obj[key] !== null) {
-          if (Array.isArray(obj[key])) {
-            res[propName] = obj[key].map((item: any) => JSON.stringify(item)).join(',');
-          } else {
-            flattenObject(obj[key], propName, res);
-          }
-        } else {
-          res[propName] = obj[key];
-        }
-      }
-      return res;
-    };
-  
+
     const flattenedData = data.map(item => flattenObject(item));
-  
+
     const workbook = XLSX.utils.book_new();
     const worksheet = XLSX.utils.json_to_sheet(flattenedData);
     XLSX.utils.book_append_sheet(workbook, worksheet, 'Sheet1');
-  
+
     XLSX.writeFile(workbook, fileName);
   };
 
@@ -46,7 +32,7 @@ export default function ProjectResources() {
     const today = new Date();
     const projectId = router.query.project;
     const formattedDate = today.toISOString().split('T')[0].replace(/-/g, '');
-    
+
     exportData(data, `${projectId}_resources_${formattedDate}.xlsx`);
   }
